+++ tests/component-plugin/src/main/resources/META-INF/mule-artifact/mule-artifact.json
@@ -7,8 +7,8 @@
       "exportedPackages": [
         "org.mule.functional.api.component",
         "org.mule.functional.api.exception",
-        "org.mule.functional.api.notification",
-        "org.mule.functional.client"
+        "org.mule.functional.api.flow",
+        "org.mule.functional.api.notification"
       ],
       "exportedResources": [
         "META-INF/spring.schemas",
